license-circuits: Adapt to new circuit caching
