+++ circuits/license/Makefile
@@ -6,10 +6,13 @@ help: ## Display this help screen
 all: ## Build the circuits
 	@cargo build --release
 
+circuits: ## Compress and store license circuit
+	@cargo run --release --features store
+
 test: ## Run the transfer circuits tests
 	@cargo test --release --features rusk-profile
 
 clippy: ## Run clippy
 	@cargo clippy --all-features --release -- -D warnings
 
-.PHONY: all test help
+.PHONY: all circuits test help
