+++ circuits/license/src/lib.rs
@@ -4,8 +4,38 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-mod error;
-mod license_circuit;
+use dusk_plonk::prelude::{Circuit, Composer, Error as PlonkError};
 
+use zk_citadel::gadgets;
+use zk_citadel::license::{CitadelProverParameters, SessionCookie};
+
+mod error;
 pub use error::Error;
-pub use license_circuit::{LicenseCircuit, ARITY, DEPTH};
+
+pub const DEPTH: usize = 17; // depth of the n-ary Merkle tree
+pub const ARITY: usize = 4; // arity of the Merkle tree
+
+#[derive(Default, Debug)]
+pub struct LicenseCircuit {
+    lpp: CitadelProverParameters<DEPTH, ARITY>,
+    sc: SessionCookie,
+}
+
+impl LicenseCircuit {
+    pub fn new(
+        lpp: CitadelProverParameters<DEPTH, ARITY>,
+        sc: SessionCookie,
+    ) -> Self {
+        Self { lpp, sc }
+    }
+}
+
+impl Circuit for LicenseCircuit {
+    fn circuit<C>(&self, composer: &mut C) -> Result<(), PlonkError>
+    where
+        C: Composer,
+    {
+        gadgets::use_license_citadel(composer, &self.lpp, &self.sc)?;
+        Ok(())
+    }
+}
